+++ Multiplayer-Roguelike.Unity/Assets/Runtime/Ecs/Core/QueryBuffer.cs
@@ -4,15 +4,15 @@ namespace Runtime.Ecs.Core
 {
     public class QueryBuffer<T> where T : IComponent
     {
-        public int Count;
+        public ushort Count;
 
-        public int[] EntityIds;
+        public ushort[] EntityIds;
 
         public T[] Components;
 
-        public QueryBuffer(int initialCapacity = 32)
+        public QueryBuffer(ushort initialCapacity = 32)
         {
-            EntityIds = new int[initialCapacity];
+            EntityIds = new ushort[initialCapacity];
             Components = new T[initialCapacity];
             Count = 0;
         }
@@ -20,17 +20,17 @@ namespace Runtime.Ecs.Core
 
     public class QueryBuffer<T1, T2> where T1 : IComponent where T2 : IComponent
     {
-        public int Count = 0;
+        public ushort Count = 0;
 
-        public int[] EntityIds;
+        public ushort[] EntityIds;
 
         public T1[] Components1;
 
         public T2[] Components2;
 
-        public QueryBuffer(int initialCapacity = 32)
+        public QueryBuffer(ushort initialCapacity = 32)
         {
-            EntityIds = new int[initialCapacity];
+            EntityIds = new ushort[initialCapacity];
             Components1 = new T1[initialCapacity];
             Components2 = new T2[initialCapacity];
             Count = 0;
@@ -42,17 +42,17 @@ namespace Runtime.Ecs.Core
         where T2 : IComponent
         where T3 : IComponent
     {
-        public int Count = 0;
+        public ushort Count = 0;
 
-        public int[] EntityIds;
+        public ushort[] EntityIds;
 
         public T1[] Components1;
         public T2[] Components2;
         public T3[] Components3;
 
-        public QueryBuffer(int initialCapacity = 32)
+        public QueryBuffer(ushort initialCapacity = 32)
         {
-            EntityIds = new int[initialCapacity];
+            EntityIds = new ushort[initialCapacity];
             Components1 = new T1[initialCapacity];
             Components2 = new T2[initialCapacity];
             Components3 = new T3[initialCapacity];
@@ -65,18 +65,18 @@ namespace Runtime.Ecs.Core
         where T3 : IComponent
         where T4 : IComponent
     {
-        public int Count = 0;
+        public ushort Count = 0;
 
-        public int[] EntityIds;
+        public ushort[] EntityIds;
 
         public T1[] Components1;
         public T2[] Components2;
         public T3[] Components3;
         public T4[] Components4;
 
-        public QueryBuffer(int initialCapacity = 32)
+        public QueryBuffer(ushort initialCapacity = 32)
         {
-            EntityIds = new int[initialCapacity];
+            EntityIds = new ushort[initialCapacity];
             Components1 = new T1[initialCapacity];
             Components2 = new T2[initialCapacity];
             Components3 = new T3[initialCapacity];
@@ -91,9 +91,9 @@ namespace Runtime.Ecs.Core
         where T4 : IComponent
         where T5 : IComponent
     {
-        public int Count = 0;
+        public ushort Count = 0;
 
-        public int[] EntityIds;
+        public ushort[] EntityIds;
 
         public T1[] Components1;
         public T2[] Components2;
@@ -101,9 +101,9 @@ namespace Runtime.Ecs.Core
         public T4[] Components4;
         public T5[] Components5;
 
-        public QueryBuffer(int initialCapacity = 32)
+        public QueryBuffer(ushort initialCapacity = 32)
         {
-            EntityIds = new int[initialCapacity];
+            EntityIds = new ushort[initialCapacity];
             Components1 = new T1[initialCapacity];
             Components2 = new T2[initialCapacity];
             Components3 = new T3[initialCapacity];
